requirements.txt
celery==4.4.2
Flask==1.1.2
lightgbm==2.2.3
pandas==1.0.3
numpy==1.18.1
python-dotenv==0.13.0
scikit_learn==0.22.2.post1
redis==3.5.0              
gunicorn==20.0.4
pymongo==3.10.1